move FastLstmBuilder to a separate file
