+++ contrib/swig/src/main/scala/edu/cmu/dynet/FastLstmBuilder.scala
@@ -1,0 +1,12 @@
+package edu.cmu.dynet
+
+class FastLstmBuilder private[dynet](private[dynet] val builder: internal.FastLSTMBuilder)
+  extends RnnBuilder(builder) {
+
+  def this() { this(new internal.FastLSTMBuilder()) }
+
+  def this(layers: Long, inputDim: Long, hiddenDim: Long, model: Model) {
+    this(new internal.FastLSTMBuilder(layers, inputDim, hiddenDim, model.model))
+  }
+
+}
